+++ notifications/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 This file documents changes to the `workbench-notifications` library, including notes on how to upgrade to new versions.
 
+## 0.3
+
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-util" % "0.3-TRAVIS-REPLACE-ME"`
+
+### Changed
+- Moved `SamModel`
+
 ## 0.2
 - upgrade cats to 1.4.0 and scala to 2.12.7
 - turn on more scalac options. upgrade scala to 2.12.11